requirements.txt
crawl4ai>=0.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
rich>=13.0.0
aiofiles>=23.0.0
